// frontend/src/app/api/model/person-dto.ts
/**
 * Coach Buddy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PhoneNumberDto } from './phone-number-dto';


export interface PersonDto { 
    id?: string;
    firstName?: string;
    lastName?: string;
    birthDate?: string;
    nickname?: string;
    notes?: string;
    phoneNumbers?: Array<PhoneNumberDto>;
    createdAt?: string;
    updatedAt?: string;
}