// frontend/src/app/api/model/update-person-request.ts
/**
 * Coach Buddy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UpdatePhoneNumberRequest } from './update-phone-number-request';


export interface UpdatePersonRequest { 
    firstName: string;
    lastName: string;
    birthDate?: string;
    nickname?: string;
    notes?: string;
    phoneNumbers?: Array<UpdatePhoneNumberRequest>;
}